refactor: 네이밍 변경
- slice 폴더 내부 파일 뒤에 slice란 변수는 불필요하다고 판단해 리팩토링
- 일부 컨벤션 통일(tagList, workspace(??List)) => (tagList, workspaceList)
