+++ client/src/components/TagList/TagList.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { fetchWorkspace } from "store/slice/workspcaeSlice";
+import { fetchWorkspace } from "store/slice/workspaceList";
 
 import Tag from "components/Tag";
 import styles from "./TagList.module.scss";
