+++ client/src/components/Workspace/Workspace.tsx
@@ -9,7 +9,7 @@ import Swal from "sweetalert2";
 import { Button } from "reactstrap";
 import AlertModal from "components/AlertModal";
 
-import { fetchWorkspace } from "store/slice/workspcaeSlice";
+import { fetchWorkspace } from "store/slice/workspaceList";
 
 import { IFolder } from "types/folder";
